# src/main/webapp/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Smart Agriculture Nutrition",
    "description": "REST API connecting smart agriculture data with personal nutrition tracking. This API bridges the gap between agricultural production and personal health outcomes, creating a complete food-to-health ecosystem.",
    "contact": {
      "name": "Smart Agriculture Team",
      "url": "https://github.com/agriculture-nutrition/api",
      "email": "agriculture@nutrition.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/SmartAgricultureNutrition/api/v1",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/personal-nutrition-trackers/{trackerId}": {
      "get": {
        "tags": ["Personal Nutrition Trackers"],
        "summary": "Get specific nutrition tracker",
        "description": "Retrieve a specific nutrition tracker with HATEOAS navigation links",
        "parameters": [
          {
            "name": "trackerId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Nutrition tracker retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalNutritionTracker"
                }
              }
            }
          },
          "404": { "description": "Nutrition tracker not found" }
        }
      },
      "put": {
        "tags": ["Personal Nutrition Trackers"],
        "summary": "Update nutrition tracker",
        "description": "Update an existing nutrition tracker with recalculated farm-to-fork scoring",
        "parameters": [
          {
            "name": "trackerId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalNutritionTracker"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nutrition tracker updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalNutritionTracker"
                }
              }
            }
          },
          "404": { "description": "Nutrition tracker not found" }
        }
      },
      "delete": {
        "tags": ["Personal Nutrition Trackers"],
        "summary": "Delete nutrition tracker",
        "description": "Remove a nutrition tracker and all associated data",
        "parameters": [
          {
            "name": "trackerId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "204": { "description": "Nutrition tracker deleted successfully" },
          "404": { "description": "Nutrition tracker not found" }
        }
      }
    },
    "/personal-nutrition-trackers/{trackerId}/meal-sources": {
      "get": {
        "tags": ["Personal Nutrition Trackers"],
        "summary": "Get meal sources for tracker",
        "description": "Retrieve meal sources with farm-to-fork tracking and local sourcing percentages",
        "parameters": [
          {
            "name": "trackerId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "local_percentage",
            "in": "query",
            "description": "Filter by minimum local percentage",
            "schema": { "type": "number", "format": "double" },
            "example": 0.8
          },
          {
            "name": "date",
            "in": "query",
            "description": "Filter by date (YYYY-MM-DD)",
            "schema": { "type": "string" },
            "example": "2024-03-15"
          }
        ],
        "responses": {
          "200": {
            "description": "Meal sources retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/MealSource" }
                }
              }
            }
          },
          "404": { "description": "Nutrition tracker not found" }
        }
      },
      "post": {
        "tags": ["Personal Nutrition Trackers"],
        "summary": "Add meal source to tracker",
        "description": "Add a new meal source with automatic local sourcing percentage and nutritional density calculation",
        "parameters": [
          {
            "name": "trackerId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MealSource" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Meal source added successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MealSource" }
              }
            }
          },
          "404": { "description": "Nutrition tracker not found" }
        }
      }
    },
...
    "/personal-nutrition-trackers/{trackerId}/bmi-analysis": {
      "get": {
        "tags": ["Personal Nutrition Trackers"],
        "summary": "Get BMI analysis for tracker",
        "description": "BMI analysis with personalized recommendations based on current health status",
        "parameters": [
          {
            "name": "trackerId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "BMI analysis completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userName": { "type": "string" },
                    "bmi": { "type": "number", "format": "double" },
                    "category": { "type": "string" },
                    "recommendations": { "type": "string" },
                    "farmToForkScore": { "type": "number", "format": "double" }
                  }
                }
              }
            }
          },
          "404": { "description": "Nutrition tracker not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PersonalNutritionTracker": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "userName": { "type": "string", "example": "John Farmer" },
          "age": { "type": "integer", "example": 34 },
          "currentBmi": {
            "type": "number",
            "format": "double",
            "example": 24.5
          },
          "healthGoals": { "type": "array", "items": { "type": "string" } },
          "farmToForkScore": {
            "type": "number",
            "format": "double",
            "example": 8.2
          }
        }
      },
...
      "MealSource": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "trackerId": { "type": "integer", "format": "int64" },
          "mealName": { "type": "string", "example": "Organic Tomato Salad" },
          "mealDate": { "type": "string", "example": "2024-03-15" },
          "ingredients": { "type": "array", "items": { "type": "string" } },
          "localSourcePercentage": {
            "type": "number",
            "format": "double",
            "example": 0.85
          },
          "nutritionalDensity": {
            "type": "number",
            "format": "double",
            "example": 7.8
          },
          "farmOrigin": {
            "type": "string",
            "example": "Local Organic Farm, Jyväskylä"
          },
          "links": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Link" }
          }
        }
      },
      "Link": {
        "type": "object",
        "properties": {
          "link": { "type": "string" },
          "rel": { "type": "string" }
        }
      }
    }
  }
}